Adding space before HTML anchor tag.
Because when we call 'create_links' method I will generate few HTML a tag based on my link. 
But, When we use CSS block level a tag, the box is showing side by side. 
In this situation, you need to add extra JS for making space. 
But, If we add the extra space in our library, It will help us to prevent the problem. 
But, If we keep the space It won't make any problem for general use.

Before space: https://i.imgur.com/ex1wZSi.png
After spae: https://i.imgur.com/YXvFYxA.png